Agregado e incializado variables que producian warnings

+++ ProyectosSDL/HolaSDL/Vector2D.h
@@ -12,8 +12,8 @@ public:
 	Vector2D() {};
 	Vector2D(int x, int y) { X = x; Y = y; };
 
-	int getX() const { return X; };
-	int getY() const { return Y; };
+	double getX() const { return X; };
+	double getY() const { return Y; };
 	void setX(int valor) { X = valor; };
 	void setY(int valor) { Y = valor; };
 	void subVectorInY(Vector2D vel) { Y -= vel.Y; };
